# docs/openapi/guesstoday-api.yaml
openapi: 3.0.3
info:
  title: GuessToday Mobile API (partial)
  description: >
    Contrat aligné sur le backend Express et les clients TypeScript
    `src/api/client.ts` et `src/api/wikiClient.ts`. Cookies `mg_session`
    (session de partie) + optionnel `Authorization: Bearer <user_session>`.
  version: 1.0.0
servers:
  - url: https://api.example.com
    description: Remplacer par l’URL réelle (ex. prod GuessToday)
...
paths:
...
  /api/friends:
    get:
      tags: [friends]
      parameters:
        - name: date
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FriendsResponse' }
...
  /api/friends/add:
    post:
      tags: [friends]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  pending:
                    type: object
                    nullable: true
                    properties:
                      id: { type: integer }
                      displayName: { type: string }
  /api/friends/accept:
    post:
      tags: [friends]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /api/friends/{userId}:
    delete:
      tags: [friends]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
components:
  schemas:
...
    FriendScore:
      type: object
      nullable: true
      properties:
        attemptsUsed: { type: integer }
        won: { type: boolean }
        completedAt: { type: string }
    FriendEntry:
      type: object
      properties:
        id: { type: integer }
        displayName: { type: string }
        streak: { type: integer }
        isMe: { type: boolean }
        scores:
          type: object
          properties:
            film: { $ref: '#/components/schemas/FriendScore' }
            series: { $ref: '#/components/schemas/FriendScore' }
            wiki: { $ref: '#/components/schemas/FriendScore' }
    PendingEntry:
      type: object
      properties:
        id: { type: integer }
        displayName: { type: string }
        direction: { type: string, enum: [incoming, outgoing] }
    FriendsResponse:
      type: object
      properties:
        date: { type: string }
        today: { type: string }
        myCode: { type: string, nullable: true }
        friends:
          type: array
          items: { $ref: '#/components/schemas/FriendEntry' }
        pending:
          type: array
          items: { $ref: '#/components/schemas/PendingEntry' }